feat: latelyPosts onClick event
